[add] optional path verification on loading
built as optional since it may be desired to load files that are currently unavailable, for example on a server that is down.
if any file (existing or not) is allowed to be loaded, then after "play" a "is_playing" command should be executed to verify server was able to fetch media.